Significant refactor of the merge processes in order to make it possible to see the differences more easily.  Specifically:
 The system now tries to only show two columns when there are differences and only shows the "auto" column when auto is not verbatim left or right.  This usually only occurs when a merge action has been offered.
In order to achieve the above, I have had to also refactor several other areas that enable it.  For example, I have changed the model definitions a little and created __get/setitems__ functions, and I have improved a lot of the type hinting etc.
Finally, this not only needed underlying logic updates, but it also needed TUI updates as well.

This is all believed to be working (so now on to the next problem area...)