I managed to introduce a bug when changing the append "-out" routines.  This is not fixed and I am not sure how I didn't see this when testing...
More significantly though I have introduced a new function that re-numbers the Finding templates. Previously this was not done at all and resulted in some clashing.  My understanding at the moment is that these IDs are just there for internal GW use and functionally mean nothing to us as users of GW.  I have subsequently removed the proposed entry points into another function to handle this numbering issue as they were not functional and wern't actually in the right place in the end anyway.